set added to the php-cs-fixer composer
