+++ src/Validation/ScriptNameValidator.php
@@ -13,9 +13,11 @@ class ScriptNameValidator
         if (!strpbrk($scriptName, '/\\')) {
             throw new IncorrectScriptNameException("Invalid script name: {$scriptName}");
         }
-        if (!preg_match('/^((\\\\?|\/?)[A-Za-z_][\w*)([\\\\\/][A-Za-z_]\w*)*$/', $scriptName)) {
+
+        if (!preg_match('/^((\\\\?|\/?)[A-Za-z_]\w*)([\\\\\/][A-Za-z_]\w*)*$/', $scriptName)) {
             throw new IncorrectScriptNameException("Invalid namespace format: {$scriptName}");
         }
+
         return true;
     }
 }
